Extract Parser and Metadata classes from parser_test file.
